fix hosting

+++ app/Http/Controllers/Admin/AsesorController.php
@@ -137,7 +137,7 @@ class AsesorController extends Controller
     {
         $user = User::findOrFail($id);
         
-        if ($user->id === auth()->id()) {
+        if ((int) $user->id === (int) auth()->id()) {
             return back()->with('error', 'Anda tidak dapat menghapus akun Anda sendiri.');
         }
